Fix: Make plot updates independent of original frequency of BB msg
Originally, the rate at which the BB was queried for latest data messages, and thus the live plot was updated, was fixed (a given cycle time of 0.1 seconds). However, the frequency at which the messages were originally logged by the BB is dependent on the frequency at which that topic (or other source) was publishing/sending these.
Since this varies, the streaming cycle was made adaptive to the original message frequency.

In order to avoid having to adjust the cycle frequency to match the frequency of each possible message, the timestamp of each (which is given by the BB) is be checked. The plot is updated only if it is different from the last message's timestamp. The cycle frequency is then fixed at some (ideally high) value.